Backend/requirements.txt
django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0
django-filter==25.1.0
cryptography==45.0.6
python-dotenv==1.1.1
psycopg2-binary==2.9.10
pillow==11.3.0
celery==5.5.3
pika==1.3.2
django-celery-results==2.6.0
amqp==5.3.1
pytz==2025.2
django-redis==6.0.0
python-dateutil==2.9.0
drf-spectacular
drf-spectacular-sidecar
gunicorn==23.0.0
django-ckeditor-5==0.2.18


# RAG & LLM
groq==0.9.0
langchain==0.2.16
langchain-groq==0.1.9
langchain-community==0.2.16
langchain-chroma==0.1.2

# Vector Store
chromadb==0.5.0
sentence-transformers==3.0.1

# PDF Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Utilities
python-decouple==3.8